Filtersets API
Filtersets are a novel way of querying spans in ENTRACE, inspired by
nixpkgs's Filesets.
Filtersets are a generalization of the old en_filter function.

A Filterset is a description of a set of span ids. For example, a
Primitive filterset is a list of ids or a range of ids; a Filter
filterset is the set of ids that match the filter predicate.

These description compose via relations such as AND, OR, NOT, etc, this allows filtersets to
have an expressive SQL-like API with a minimal backend implementation.
In a way, filtersets encode dataflow itself.

The other defining feature of a Filterset is that it can be
*materialized* (turned into a Vec<id>) freely at any time, which aids
debugging, and makes it trivial to switch between Filtersets and
"traditional" lua when desired.

The performance speedup (1.3-2x) of filtersets primarily comes from the fact that
via filtersets, many allocations in the Lua evaluator can be avoided;
and that the Filterset evaluator includes a "normalization" step, which
tries to rewrite the filterset to the fastest possible query.

Filtersets have a long way to go still. The filterset evaluator itself
is very generic on purpose; it assumes nothing about ENTRACE itself.
In the future, it can (and should) be extracted from this project, as it
may be useful in other software as well.

There is still a long way to go with filtersets, many possible
optmizations (e.g. introducing a DNF type in the evaluator to reduce
scans in queries having OR), a BlackBox type that materializes its
source without rewriting; but they are already powerful enough to speed
up queries; and to provide a nicer syntax.