Filtersets V2
Simplifies filtersets. The primary filter type is now a RelDnf (the ids
of spans matching a Disjunctive Normal Form over a source).
This makes rewrite rules simpler (many -> one (intersection of DNFS)),
and also results in better performance, since we can merge many kinds of
nodes into one DNF, which corresponds to only a single scan on the data.